Move menu to own class (#21)
* Added the menu class and made an if statement after Login() to check if user is admin or not. Also wrote out a different Login() with authenticate()

* Created a more ordered filestructure with folders and moved relevant files to each. Added a abstract menu and inheriting classes AdminMenu and CustomerMenu

* Changed the menu option for customer with MarkdownUtils that uses markdown. Added a while loop to let program to run after user logged out.

* Uncommented the UnlockUser. Made a mistake commenting it out

* Took away CustomerMenuLoan from bank.cs and chnaged name of a method in CustomerMenu.cs

* Fixed some errors that happened after updating from develop branch.
