+++ RebelAllianceBank/Accounts/CardAccount.cs
@@ -1,18 +1,20 @@
 ﻿using RebelAllianceBank.Interfaces;
-namespace RebelAllianceBank.Classes
+namespace RebelAllianceBank.Accounts
 {
-    public class SavingsAccount : IBankAccount
+    public class CardAccount : IBankAccount
     {
         public int ID { get; set; }
         public string UserId { get; set; }
-        public int AccountType { get; set; } = 1;
+        public int AccountType { get; set; } = 0;
         public string AccountName { get; set; }
         public decimal Balance { get; set; }
         public string AccountCurrency { get; set; }
         public decimal IntrestRate { get; set; }
-        public SavingsAccount() { }
-        public SavingsAccount(string accountName, decimal balance, string accountCurrency, decimal intrestRate = 0)
+        public CardAccount() { }
+        public CardAccount(string userId, int accountType, string accountName, decimal balance, string accountCurrency, decimal intrestRate = 0)
         {
+            UserId = userId;
+            AccountType = accountType;
             AccountName = accountName;
             Balance = balance;
             AccountCurrency = accountCurrency;
