+++ RebelAllianceBank/Accounts/SavingsAccount.cs
@@ -1,20 +1,18 @@
 ﻿using RebelAllianceBank.Interfaces;
-namespace RebelAllianceBank.Classes
+namespace RebelAllianceBank.Accounts
 {
-    public class CardAccount : IBankAccount
+    public class SavingsAccount : IBankAccount
     {
         public int ID { get; set; }
         public string UserId { get; set; }
-        public int AccountType { get; set; } = 0;
+        public int AccountType { get; set; } = 1;
         public string AccountName { get; set; }
         public decimal Balance { get; set; }
         public string AccountCurrency { get; set; }
         public decimal IntrestRate { get; set; }
-        public CardAccount() { }
-        public CardAccount(string userId, int accountType, string accountName, decimal balance, string accountCurrency, decimal intrestRate = 0)
+        public SavingsAccount() { }
+        public SavingsAccount(string accountName, decimal balance, string accountCurrency, decimal intrestRate = 0)
         {
-            UserId = userId;
-            AccountType = accountType;
             AccountName = accountName;
             Balance = balance;
             AccountCurrency = accountCurrency;
