+++ RebelAllianceBank/Menu/Menu.cs
@@ -1,0 +1,15 @@
+﻿using RebelAllianceBank.Interfaces;
+using RebelAllianceBank.utils;
+
+namespace RebelAllianceBank.Menu
+{
+    public abstract class Menu
+    {
+        protected IUser CurrentUser { get; }
+        protected Menu(IUser currentUser)
+        {
+            CurrentUser = currentUser;
+        }
+        public abstract void Show();
+    }
+}
